Added double-checking of time labels and conflict checking on import
Valid time labels are contained within a separate file now for error checking. This is important because I feel it would be very easy to have a typo in the time.

Added code to provide a warning when importing a schedule if a faculty member has a conflict.

Polished export formatting a little bit more.