Added ability to correctly handle multi-section courses and small update for format of input files
Create a custom classes file (changed the desired format slightly)

Added the ability to include comments within any of the input files for GrandAlpha

Added the notion of courses that are multi-section and those that are not. Added a function to determine by name what kind a course is.

Changed the penalty function so that multi-section courses do not give penalties under any circumstances.