Refactored time detail info
The day and time information for the actual strings that need to appear in the final report based on a given time code has been removed from the code and placed into the ValidTimes information file. It is now much easier to update and maintain the list of valid times because the detailed information is right next to it.