New initialization and report formats
Changing the way that Grand Alpha is initialized. Instead of customized files that define course offerings, possible times, and faculty assignments, this information is now input in a format that is as close as possible to the format used for exporting reports. This simplifies the back and forth between how I communicate with the other faculty and what the code needs to run. A sample input file is provided here to show how this works. Most information in the report format is ignored during initialization and import. It is regenerated during export.

Along with the new format for initialization, made slight changes to the detailed export format. Also created the ability to read in a schedule from the detailed export format. Long term the simple import/export format should be removed.

Detailed export now takes a flag to determine whether cross listings will be added.