Added feature to define conflicts in clusters
Conflicts can now be defined by listing a set of courses that should not conflict with each other instead of having to list all the conflicts pairwise. Needed to change the format of the Conflicts file slightly to accommodate the possibility having several courses listed.

Changed behavior so that listing conflicts mutliple times does not accumulate penalty weight, which is how I have been thinking about it in the past. I have been careful in the past to not list conflicts mutliple times, which is no longer necessary.