Update Fall 2020
This updates based on changes during the production of the schedule for Spring 2021.

Updated primary format used for input/output to include codes for mode of delivery and hours of teaching load.

Improved initialization of CourseTimes class so that a single line as a string does not need to be produced. Also added various information during initialization so that the entire input file can be reproduced on export with nothing lost.

Added feature to ignore cross listings on init and import, meaning that cross listings can now always be included on export without complicating the use of the file later in the process.