Removed need for manually creating course codes
Refactored GrandAlpha so that it can import files using the registrar's course code format. A course code will be generated from this that is consistent with the normal course code format, but now it will not need to be in the spreadsheet.

Changed some defaults so that detailed schedules are the only ones that are imported and exported.

Little bit of cleanup